Implemented pixel rendering.
Should be fairly inefficient but feels fast due to caching.

+++ src/model/buffer_handling.rs
@@ -139,7 +139,7 @@ impl Buffer {
         std::mem::replace( &mut self.overlay_layer, None)
     }
 
-    pub fn get_font_scanline(&self, ch: u8, y: usize) -> u32
+    pub fn get_font_scanline(&self, ch: u16, y: usize) -> u32
     {
         self.font.get_scanline(ch, y)
     }
